refactor(api): move update-certification-center usecase
in organizational-entities
